requirements.txt
scikit-learn==1.5.2
numpy
joblib
pillow
matplotlib
seaborn
